requirements.txt
allure-behave==2.6.2
allure-python-commons==2.6.2
Appium-Python-Client==0.48
asn1crypto==0.24.0
attrs==19.1.0
behave==1.2.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Click==7.0
importlib-metadata==0.19
idna==2.8
parse==1.12.0
parse-type==0.4.2
pluggy==0.9.0
public==2019.4.13
pycparser==2.19
pymssql==2.1.4
PyYAML==5.1
query-string==2019.4.13
requests==2.21.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.3
zipp==0.5.2
